api/requirements.txt
uwsgi~=2.0
Django~=1.11
mysqlclient~=1.3
djangorestframework~=3.6
requests~=2.19